requirements.txt
django==2.1
django-extensions==2.1.2
django-cors-headers==2.4.0
django-storages==1.6.6
django-smtp-ssl==1.0
pymysql==0.9.2
werkzeug==0.14.1
boto3==1.8.3
botocore==1.11.3
django-bootstrap-form==3.4
customerio==0.4
shipstation==0.1.2
django-recaptcha2==1.3.0
paramiko==2.4.1
ipython==6.5.0
django-bootstrap3==10.0.1
unicodecsv==0.14.1
django-admin-tools==0.8.1
django-redis==4.9.0
django-bulk-update==2.2.0
django-debug-toolbar==1.9.1
xhtml2pdf==0.2.2
django-anymail[mailgun]==4.1
vultr==1.0.1
django-filter==2.0.0
djangorestframework==3.9.0
markdown==3.0.1
coreapi==2.3.3
